requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyascii==0.3.2
asgiref==3.8.1
Authlib==1.5.2
beautifulsoup4==4.13.3
billiard==4.2.1
bleach==3.3.1
blinker==1.9.0
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==44.0.2
defusedxml==0.7.1
Django==5.1.7
django-appconf==1.1.0
django-axes==7.0.2
django-ckeditor-5==0.2.17
django-cleanup==9.0.0
django-compressor==4.5.1
django-filter==24.3
django-formtools==2.5.1
django-js-asset==3.1.2
django-modelcluster==6.4
django-permissionedforms==0.1
django-ratelimit==4.1.0
django-recaptcha==3.0.0
django-stubs-ext==5.1.3
django-taggit==6.1.0
django-tasks==0.6.1
django-treebeard==4.7.1
djangorestframework==3.15.2
dparse==0.6.4
draftjs_exporter==5.1.0
et_xmlfile==2.0.0
filelock==3.16.1
filetype==1.2.0
gitdb==4.0.12
GitPython==3.1.44
googletrans==4.0.0rc1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
itsdangerous==2.2.0
jalali_core==1.0.0
jdatetime==5.1.0
Jinja2==3.1.5
joblib==1.4.2
kombu==5.5.0
laces==0.1.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
nltk==3.9.1
openpyxl==3.1.5
packaging==24.2
pbr==6.1.1
pillow==11.1.0
pillow_heif==0.21.0
prompt_toolkit==3.0.50
psutil==6.1.1
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.19.1
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
qrcode==7.4.2
rcssmin==1.1.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
rich==14.0.0
rjsmin==1.2.2
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
safety-schemas==0.0.11
sentry-sdk==2.24.1
setuptools==78.1.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
stevedore==5.4.1
telepath==0.3.1
tinycss2==1.1.1
tqdm==4.67.1
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
Willow==1.9.0
